+++ src/test/java/com/vetrikos/inventory/system/mapper/ItemMapperTest.java
@@ -1,7 +1,5 @@
 package com.vetrikos.inventory.system.mapper;
 
-import static org.assertj.core.api.Assertions.assertThat;
-
 import com.vetrikos.inventory.system.entity.Item;
 import com.vetrikos.inventory.system.model.WarehouseItemRequestRestDTO;
 import org.junit.jupiter.api.Test;
@@ -10,31 +8,33 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.test.context.ContextConfiguration;
 import org.springframework.test.context.junit.jupiter.SpringExtension;
 
+import static org.assertj.core.api.Assertions.assertThat;
+
 @ExtendWith(SpringExtension.class)
 @ContextConfiguration(classes = {
-    ItemMapperImpl.class,
+        ItemMapperImpl.class,
 })
 class ItemMapperTest {
 
-  @Autowired
-  private ItemMapper itemMapper;
+    @Autowired
+    private ItemMapper itemMapper;
 
 
-  @Test
-  void warehouseItemRequestRestDTOToItem() {
-    int size = 500;
-    String name = "Apple";
-    int quantity = 30;
-    WarehouseItemRequestRestDTO requestRestDTO = new WarehouseItemRequestRestDTO(name, size, quantity);
+    @Test
+    void warehouseItemRequestRestDTOToItem() {
+        long size = 500;
+        String name = "Apple";
+        long quantity = 30;
+        WarehouseItemRequestRestDTO requestRestDTO = new WarehouseItemRequestRestDTO(name, size, quantity);
 
-    Item expectedItem = Item.builder()
-        .name(name)
-        .size(size)
-        .build();
+        Item expectedItem = Item.builder()
+                .name(name)
+                .size(size)
+                .build();
 
-    Item result = itemMapper.warehouseItemRequestRestDTOToItem(requestRestDTO);
+        Item result = itemMapper.warehouseItemRequestRestDTOToItem(requestRestDTO);
 
-    assertThat(result).isNotNull()
-        .isEqualTo(expectedItem);
-  }
+        assertThat(result).isNotNull()
+                .isEqualTo(expectedItem);
+    }
 }
